added comprehensive ui tests for backgrounding
